Refactor - some more synthax details

+++ BMI Calculator/BMI Calculator.js
@@ -2,8 +2,8 @@
 
 let WeightInput = prompt("Type your weight in kg here")
 let HeightInput = prompt("Type your height in meters here")
-let Weight = parseInt(WeightInput)
-let Height = parseInt(HeightInput)
+let Weight = parseFloat(WeightInput)
+let Height = parseFloat(HeightInput)
 let BMI = Weight / (Height**2)
 if (BMI<=18.5) {
     // alert(BMI);
